address third review/feedback changes
