+++ packages/api/src/channels/channels.controller.ts
@@ -1,13 +1,53 @@
-import { VideosRepository } from '@joystream/ytube'
-import { Body, Controller, Get, NotFoundException, Param, Put } from '@nestjs/common'
+import { ChannelsRepository, IYoutubeClient, UsersRepository, VideosRepository } from '@joystream/ytube'
+import { BadRequestException, Body, Controller, Get, Inject, NotFoundException, Param, Post, Put } from '@nestjs/common'
 import { ApiBody, ApiOperation, ApiResponse, ApiTags } from '@nestjs/swagger'
-import { ChannelDto, UpdateChannelDto, VideoDto } from '../dtos'
+import { Channel, User } from '@youtube-sync/domain'
+import { ChannelDto, SaveChannelRequest, SaveChannelResponse, UpdateChannelDto, UserDto, VideoDto } from '../dtos'
 import { ChannelsService } from './channels.service'
 
 @Controller('channels')
 @ApiTags('channels')
 export class ChannelsController {
-  constructor(private channelsService: ChannelsService, private videosRepository: VideosRepository) {}
+  constructor(
+    @Inject('youtube') private youtube: IYoutubeClient,
+    private channelsService: ChannelsService,
+    private channelsRepository: ChannelsRepository,
+    private usersRepository: UsersRepository,
+    private videosRepository: VideosRepository
+  ) {}
+
+  @ApiOperation({
+    description: `Creates user from the supplied google authorization code and fetches
+     user's channel and if it satisfies YPP induction criteria it saves the record`,
+  })
+  @ApiBody({ type: SaveChannelRequest })
+  @ApiResponse({ type: SaveChannelResponse })
+  @Post('/verify-and-save')
+  async addVerifiedChannel(
+    @Body() { authorizationCode, userId, joystreamChannelId, referrerChannelId, email }: SaveChannelRequest
+  ): Promise<SaveChannelResponse> {
+    try {
+      // get user from userId
+      const user = await this.usersRepository.get(userId)
+
+      // ensure request's authorization code matches the user's authorization code
+      if (user.authorizationCode !== authorizationCode) {
+        throw new Error('Invalid request author. Permission denied.')
+      }
+
+      // get channel from user
+      const [channel] = await this.youtube.getChannels(user)
+
+      // save user and channel
+      await this.saveUserAndChannel({ ...user, email }, { ...channel, joystreamChannelId, referrerChannelId, email })
+
+      // return user and channel
+      return new SaveChannelResponse(new UserDto(user), new ChannelDto(channel))
+    } catch (error) {
+      const message = error instanceof Error ? error.message : error
+      throw new BadRequestException(message)
+    }
+  }
 
   @Get(':joystreamChannelId')
   @ApiOperation({ description: 'Retrieves channel by joystreamChannelId' })
@@ -50,5 +90,13 @@ export class ChannelsController {
   async getVideo(@Param('id') id: string, @Param('videoId') videoId: string) {
     const result = await this.videosRepository.get(id, videoId)
     return result
+  }
+
+  private async saveUserAndChannel(user: User, channel: Channel) {
+    // save user
+    await this.usersRepository.save(user)
+
+    // save channel
+    return await this.channelsRepository.save(channel)
   }
 }
